Fix security bugs

+++ src/app/components/reusables/images/image-preview-modal/image-preview-modal.component.ts
@@ -1,8 +1,12 @@
-import { DialogRef } from '@angular/cdk/dialog';
-import { Component, Inject, OnInit } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { Component, Inject } from '@angular/core';
+import {
+    MAT_DIALOG_DATA,
+    MatDialog,
+    MatDialogRef,
+} from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { RestaurantsService } from 'src/app/services/restaurants.service';
+import { ConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
 
 @Component({
     selector: 'image-preview-modal',
@@ -16,6 +20,7 @@ export class ImagePreviewModalComponent {
 
     constructor(
         @Inject(MAT_DIALOG_DATA) public data: ImagePreviewModalParams,
+        private dialog: MatDialog,
         private restaurantService: RestaurantsService,
         private notificationService: MatSnackBar,
         public dialogRef: MatDialogRef<ImagePreviewModalComponent>,
@@ -26,26 +31,38 @@ export class ImagePreviewModalComponent {
     }
 
     deleteImage() {
-        this.restaurantService
-            .deleteImage(this.restaurantId, this.imageUrl)
-            .subscribe({
-                next: (res) => {
-                    this.dialogRef.close(this.imageUrl);
+        this.dialog
+            .open(ConfirmDialogComponent, {
+                data: {
+                    title: 'Delete Image',
+                    message: 'Are you sure you want to delete this image?',
+                },
+            })
+            .afterClosed()
+            .subscribe((result) => {
+                if (result) {
+                    this.restaurantService
+                        .deleteImage(this.restaurantId, this.imageUrl)
+                        .subscribe({
+                            next: (res) => {
+                                this.dialogRef.close(this.imageUrl);
 
-                    this.notificationService.open(
-                        'Image successfully deleted!',
-                        'Ok',
-                        {
-                            duration: 3000,
-                        },
-                    );
-                },
-                error: (err: any) => {
-                    this.notificationService.open(
-                        `Failed to delete the image: ${err.message}`,
-                        'Oops',
-                    );
-                },
+                                this.notificationService.open(
+                                    'Image successfully deleted!',
+                                    'Ok',
+                                    {
+                                        duration: 3000,
+                                    },
+                                );
+                            },
+                            error: (err: any) => {
+                                this.notificationService.open(
+                                    `Failed to delete the image: ${err.message}`,
+                                    'Oops',
+                                );
+                            },
+                        });
+                }
             });
     }
 }
